Added the Inovis meddp-515 XGA display configuration with source patches.
To get the Inovis meddp-515 diplay to work with the controller system,
several changes where made. The pixel clock couldn't be set to a proper
value before the patch, some possible NULL-pointer exceptions are fixed,
pixel clock is inverted. Also the device tree was changed to use the
correct display type.

In addition a framebuffer testsoftware was activated within buildroot to
test the display.