Added support for the additional serial ports.
To bring the three additional serial ports up, two fixes are needed. The
one corrects the data pins of the mmc0 controller to four, as one of the
higher four data pins conflicts with the serial port. The other one
removes the LED from a xplained board devicetree file which is included
for this board, as this pin also conflicts with a pin of one serial port.